+++ src/app.js
@@ -1,19 +1,11 @@
 import * as React from 'react'
 import {useAuth} from './context/auth-context'
-import {FullPageSpinner} from './components/lib'
-
-const AuthenticatedApp = React.lazy(() =>
-  import(/* webpackPrefetch: true */ './authenticated-app'),
-)
-const UnauthenticatedApp = React.lazy(() => import('./unauthenticated-app'))
+import {AuthenticatedApp} from './authenticated-app'
+import {UnauthenticatedApp} from './unauthenticated-app'
 
 function App() {
   const {user} = useAuth()
-  return (
-    <React.Suspense fallback={<FullPageSpinner />}>
-      {user ? <AuthenticatedApp /> : <UnauthenticatedApp />}
-    </React.Suspense>
-  )
+  return user ? <AuthenticatedApp /> : <UnauthenticatedApp />
 }
 
 export {App}
