+++ src/authenticated-app.js
@@ -157,4 +157,4 @@ function AppRoutes() {
   )
 }
 
-export default AuthenticatedApp
+export {AuthenticatedApp}
