+++ src/components/list-item-list.js
@@ -4,7 +4,6 @@ import {jsx} from '@emotion/core'
 import {useListItems} from 'utils/list-items'
 import {BookListUL} from './lib'
 import {BookRow} from './book-row'
-import {Profiler} from './profiler'
 
 function ListItemList({filterListItems, noListItems, noFilteredListItems}) {
   const listItems = useListItems()
@@ -23,18 +22,13 @@ function ListItemList({filterListItems, noListItems, noFilteredListItems}) {
   }
 
   return (
-    <Profiler
-      id="List Item List"
-      metadata={{listItemCount: filteredListItems.length}}
-    >
-      <BookListUL>
-        {filteredListItems.map(listItem => (
-          <li key={listItem.id} aria-label={listItem.book.title}>
-            <BookRow book={listItem.book} />
-          </li>
-        ))}
-      </BookListUL>
-    </Profiler>
+    <BookListUL>
+      {filteredListItems.map(listItem => (
+        <li key={listItem.id}>
+          <BookRow book={listItem.book} />
+        </li>
+      ))}
+    </BookListUL>
   )
 }
