+++ src/components/modal.extra-1.js
@@ -1,0 +1,38 @@
+/** @jsx jsx */
+import {jsx} from '@emotion/core'
+
+import * as React from 'react'
+import {Dialog} from './lib'
+
+const callAll = (...fns) => (...args) => fns.forEach(fn => fn && fn(...args))
+
+const ModalContext = React.createContext()
+
+function Modal(props) {
+  const [isOpen, setIsOpen] = React.useState(false)
+
+  return <ModalContext.Provider value={[isOpen, setIsOpen]} {...props} />
+}
+
+function ModalDismissButton({children: child}) {
+  const [, setIsOpen] = React.useContext(ModalContext)
+  return React.cloneElement(child, {
+    onClick: callAll(() => setIsOpen(false), child.props.onClick),
+  })
+}
+
+function ModalOpenButton({children: child}) {
+  const [, setIsOpen] = React.useContext(ModalContext)
+  return React.cloneElement(child, {
+    onClick: callAll(() => setIsOpen(true), child.props.onClick),
+  })
+}
+
+function ModalContents(props) {
+  const [isOpen, setIsOpen] = React.useContext(ModalContext)
+  return (
+    <Dialog isOpen={isOpen} onDismiss={() => setIsOpen(false)} {...props} />
+  )
+}
+
+export {Modal, ModalDismissButton, ModalOpenButton, ModalContents}
