+++ src/components/modal.js
@@ -1,59 +1,9 @@
-/** @jsx jsx */
-import {jsx} from '@emotion/core'
+export * from './modal.final'
 
-import * as React from 'react'
-import VisuallyHidden from '@reach/visually-hidden'
-import {Dialog, CircleButton} from './lib'
+// export * from './modal.exercise'
 
-const callAll =
-  (...fns) =>
-  (...args) =>
-    fns.forEach(fn => fn && fn(...args))
+// 💯 Add `callAll`
+// export * from './modal.extra-1'
 
-const ModalContext = React.createContext()
-
-function Modal(props) {
-  const [isOpen, setIsOpen] = React.useState(false)
-
-  return <ModalContext.Provider value={[isOpen, setIsOpen]} {...props} />
-}
-
-function ModalDismissButton({children: child}) {
-  const [, setIsOpen] = React.useContext(ModalContext)
-  return React.cloneElement(child, {
-    onClick: callAll(() => setIsOpen(false), child.props.onClick),
-  })
-}
-
-function ModalOpenButton({children: child}) {
-  const [, setIsOpen] = React.useContext(ModalContext)
-  return React.cloneElement(child, {
-    onClick: callAll(() => setIsOpen(true), child.props.onClick),
-  })
-}
-
-function ModalContentsBase(props) {
-  const [isOpen, setIsOpen] = React.useContext(ModalContext)
-  return (
-    <Dialog isOpen={isOpen} onDismiss={() => setIsOpen(false)} {...props} />
-  )
-}
-
-function ModalContents({title, children, ...props}) {
-  return (
-    <ModalContentsBase {...props}>
-      <div css={{display: 'flex', justifyContent: 'flex-end'}}>
-        <ModalDismissButton>
-          <CircleButton>
-            <VisuallyHidden>Close</VisuallyHidden>
-            <span aria-hidden>×</span>
-          </CircleButton>
-        </ModalDismissButton>
-      </div>
-      <h3 css={{textAlign: 'center', fontSize: '2em'}}>{title}</h3>
-      {children}
-    </ModalContentsBase>
-  )
-}
-
-export {Modal, ModalDismissButton, ModalOpenButton, ModalContents}
+// 💯 Create ModalContentsBase
+// export * from './modal.extra-2'
