+++ src/index.js
@@ -2,16 +2,18 @@ import {loadDevTools} from './dev-tools/load'
 import './bootstrap'
 import * as React from 'react'
 import {createRoot} from 'react-dom/client'
-import {Profiler} from 'components/profiler'
 import {App} from './app'
 import {AppProviders} from './context'
 
+// ignore the rootRef in this file. I'm just doing it here to make
+// the tests I write to check your work easier.
+export const rootRef = {}
 loadDevTools(() => {
-  createRoot(document.getElementById('root')).render(
-    <Profiler id="App Root" phases={['mount']}>
-      <AppProviders>
-        <App />
-      </AppProviders>
-    </Profiler>,
+  const root = createRoot(document.getElementById('root'))
+  root.render(
+    <AppProviders>
+      <App />
+    </AppProviders>,
   )
+  rootRef.current = root
 })
