+++ src/screens/book.js
@@ -13,7 +13,6 @@ import * as mq from 'styles/media-queries'
 import * as colors from 'styles/colors'
 import {Spinner, Textarea, ErrorMessage} from 'components/lib'
 import {Rating} from 'components/rating'
-import {Profiler} from 'components/profiler'
 import {StatusButtons} from 'components/status-buttons'
 
 function BookScreen() {
@@ -24,63 +23,59 @@ function BookScreen() {
   const {title, author, coverImageUrl, publisher, synopsis} = book
 
   return (
-    <Profiler id="Book Screen" metadata={{bookId, listItemId: listItem?.id}}>
-      <div>
-        <div
-          css={{
-            display: 'grid',
-            gridTemplateColumns: '1fr 2fr',
-            gridGap: '2em',
-            marginBottom: '1em',
-            [mq.small]: {
-              display: 'flex',
-              flexDirection: 'column',
-            },
-          }}
-        >
-          <img
-            src={coverImageUrl}
-            alt={`${title} book cover`}
-            css={{width: '100%', maxWidth: '14rem'}}
-          />
-          <div>
-            <div css={{display: 'flex', position: 'relative'}}>
-              <div css={{flex: 1, justifyContent: 'space-between'}}>
-                <h1>{title}</h1>
-                <div>
-                  <i>{author}</i>
-                  <span css={{marginRight: 6, marginLeft: 6}}>|</span>
-                  <i>{publisher}</i>
-                </div>
-              </div>
-              <div
-                css={{
-                  right: 0,
-                  color: colors.gray80,
-                  display: 'flex',
-                  flexDirection: 'column',
-                  justifyContent: 'space-around',
-                  minHeight: 100,
-                }}
-              >
-                {book.loadingBook ? null : <StatusButtons book={book} />}
+    <div>
+      <div
+        css={{
+          display: 'grid',
+          gridTemplateColumns: '1fr 2fr',
+          gridGap: '2em',
+          marginBottom: '1em',
+          [mq.small]: {
+            display: 'flex',
+            flexDirection: 'column',
+          },
+        }}
+      >
+        <img
+          src={coverImageUrl}
+          alt={`${title} book cover`}
+          css={{width: '100%', maxWidth: '14rem'}}
+        />
+        <div>
+          <div css={{display: 'flex', position: 'relative'}}>
+            <div css={{flex: 1, justifyContent: 'space-between'}}>
+              <h1>{title}</h1>
+              <div>
+                <i>{author}</i>
+                <span css={{marginRight: 6, marginLeft: 6}}>|</span>
+                <i>{publisher}</i>
               </div>
             </div>
-            <div css={{marginTop: 10, minHeight: 46}}>
-              {listItem?.finishDate ? <Rating listItem={listItem} /> : null}
-              {listItem ? <ListItemTimeframe listItem={listItem} /> : null}
+            <div
+              css={{
+                right: 0,
+                color: colors.gray80,
+                display: 'flex',
+                flexDirection: 'column',
+                justifyContent: 'space-around',
+                minHeight: 100,
+              }}
+            >
+              {book.loadingBook ? null : <StatusButtons book={book} />}
             </div>
-            <br />
-            <p css={{whiteSpace: 'break-spaces', display: 'block'}}>
-              {synopsis}
-            </p>
           </div>
+          <div css={{marginTop: 10, minHeight: 46}}>
+            {listItem?.finishDate ? <Rating listItem={listItem} /> : null}
+            {listItem ? <ListItemTimeframe listItem={listItem} /> : null}
+          </div>
+          <br />
+          <p>{synopsis}</p>
         </div>
-        {!book.loadingBook && listItem ? (
-          <NotesTextarea listItem={listItem} />
-        ) : null}
       </div>
-    </Profiler>
+      {!book.loadingBook && listItem ? (
+        <NotesTextarea listItem={listItem} />
+      ) : null}
+    </div>
   )
 }
