+++ src/screens/discover.js
@@ -8,7 +8,6 @@ import * as colors from 'styles/colors'
 import {useBookSearch, useRefetchBookSearchQuery} from 'utils/books'
 import {BookRow} from 'components/book-row'
 import {BookListUL, Spinner, Input} from 'components/lib'
-import {Profiler} from 'components/profiler'
 
 function DiscoverBooksScreen() {
   const [query, setQuery] = React.useState('')
@@ -85,18 +84,13 @@ function DiscoverBooksScreen() {
           </div>
         )}
         {books.length ? (
-          <Profiler
-            id="Discover Books Screen Book List"
-            metadata={{query, bookCount: books.length}}
-          >
-            <BookListUL css={{marginTop: 20}}>
-              {books.map(book => (
-                <li key={book.id} aria-label={book.title}>
-                  <BookRow key={book.id} book={book} />
-                </li>
-              ))}
-            </BookListUL>
-          </Profiler>
+          <BookListUL css={{marginTop: 20}}>
+            {books.map(book => (
+              <li key={book.id} aria-label={book.title}>
+                <BookRow key={book.id} book={book} />
+              </li>
+            ))}
+          </BookListUL>
         ) : queried ? (
           <div css={{marginTop: 20, fontSize: '1.2em', textAlign: 'center'}}>
             {isLoading ? (
